Rename EventSpanAttributes -> EventAttributeSources
